Reorganize symbol and data_types sub-crates to clean up the top-level
